+++ Source/Services/Profiles/HolidayContractProfile.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using DbEntites = DsuDev.BusinessDays.DataAccess.Entites;
+using DbModels = DsuDev.BusinessDays.DataAccess.Models;
 using DomainEntities = DsuDev.BusinessDays.Domain.Entities;
 
 namespace DsuDev.BusinessDays.Services.Profiles
@@ -14,7 +14,7 @@ namespace DsuDev.BusinessDays.Services.Profiles
         public void CreateHolidayMapping()
         {
 
-            this.CreateMap<DbEntites.Holiday, DomainEntities.Holiday>()
+            this.CreateMap<DbModels.Holiday, DomainEntities.Holiday>()
                 .ForMember(dest => dest.Id, op => op.MapFrom(src => src.Id))
                 .ForMember(dest => dest.Name, op => op.MapFrom(src => src.Name))
                 .ForMember(dest => dest.Description, op => op.MapFrom(src => src.Description))
@@ -24,7 +24,7 @@ namespace DsuDev.BusinessDays.Services.Profiles
                 .ForAllOtherMembers(dest => dest.Ignore());
 
 
-            this.CreateMap<DomainEntities.Holiday, DbEntites.Holiday>()
+            this.CreateMap<DomainEntities.Holiday, DbModels.Holiday>()
                 .ForMember(dest => dest.Id, op => op.MapFrom(src => src.Id))
                 .ForMember(dest => dest.Name, op => op.MapFrom(src => src.Name))
                 .ForMember(dest => dest.Description, op => op.MapFrom(src => src.Description))
